Incluyendo Cambios JUNIT MOKITO

+++ ProyectoLuisAcuna/src/main/java/com/LAAR/AFP/Bootcamp/repository/IClientRepository.java
@@ -11,7 +11,7 @@ import com.LAAR.AFP.Bootcamp.entities.Client;
 import java.util.List;
 
 @Repository
-public interface IClientRepository extends JpaRepository<Client, Integer> {
+public interface IClientRepository extends JpaRepository<Client, Long> {
 
     @Query(value = "SELECT * FROM clients WHERE AFP = :AFP" ,nativeQuery = true)
     public List<Client> getClientForAFP(@Param("AFP") String AFP);
